Split KeeperTCPHandlerFactory into module part
